requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.1
python-multipart>=0.0.6
requests>=2.31.0
google-generativeai>=0.3.0
# Kestra client (optional)
# kestra>=0.1.0
# Oumi RL training (optional)
# oumi>=0.1.0
# torch>=2.0.0
